requirements.txt
torch
tqdm
whisper
pysubs2
google-generativeai
openai